README.md
An Angular app that enables a user to search tweets on Twitter by entering a keyword.You can query the Twitter search API via an 
ajax GET request and grab the resulting JSON response and display it through data binding as the search result. 
The search results should mimic a Twitter stream.